+++ src/Celerity/Collections/HashCachingDictionary.cs
@@ -570,16 +570,7 @@ public class HashCachingDictionary<TKey, TValue, THasher>
     /// <exception cref="ArgumentException"><paramref name="array"/> has insufficient space.</exception>
     public void CopyTo(KeyValuePair<TKey, TValue?>[] array, int arrayIndex)
     {
-        ArgumentNullException.ThrowIfNull(array);
-        if (arrayIndex < 0)
-            throw new ArgumentOutOfRangeException(nameof(arrayIndex), arrayIndex,
-                "Array index must be non-negative.");
-        if (arrayIndex > array.Length)
-            throw new ArgumentOutOfRangeException(nameof(arrayIndex), arrayIndex,
-                "Array index is beyond the end of the destination array.");
-        if (array.Length - arrayIndex < _count)
-            throw new ArgumentException(
-                "The destination array has insufficient space for the entries.", nameof(array));
+        CopyToGuard.Validate(array, arrayIndex, _count, CopyToGuard.EntriesMessage);
 
         int i = arrayIndex;
         foreach (KeyValuePair<TKey, TValue?> entry in this)
@@ -741,16 +732,7 @@ public class HashCachingDictionary<TKey, TValue, THasher>
         /// <exception cref="ArgumentException"><paramref name="array"/> has insufficient space.</exception>
         public void CopyTo(TKey[] array, int arrayIndex)
         {
-            ArgumentNullException.ThrowIfNull(array);
-            if (arrayIndex < 0)
-                throw new ArgumentOutOfRangeException(nameof(arrayIndex), arrayIndex,
-                    "Array index must be non-negative.");
-            if (arrayIndex > array.Length)
-                throw new ArgumentOutOfRangeException(nameof(arrayIndex), arrayIndex,
-                    "Array index is beyond the end of the destination array.");
-            if (array.Length - arrayIndex < _dict._count)
-                throw new ArgumentException(
-                    "The destination array has insufficient space for the keys.", nameof(array));
+            CopyToGuard.Validate(array, arrayIndex, _dict._count, CopyToGuard.KeysMessage);
 
             int i = arrayIndex;
             foreach (KeyValuePair<TKey, TValue?> entry in _dict)
@@ -840,16 +822,7 @@ public class HashCachingDictionary<TKey, TValue, THasher>
         /// <exception cref="ArgumentException"><paramref name="array"/> has insufficient space.</exception>
         public void CopyTo(TValue?[] array, int arrayIndex)
         {
-            ArgumentNullException.ThrowIfNull(array);
-            if (arrayIndex < 0)
-                throw new ArgumentOutOfRangeException(nameof(arrayIndex), arrayIndex,
-                    "Array index must be non-negative.");
-            if (arrayIndex > array.Length)
-                throw new ArgumentOutOfRangeException(nameof(arrayIndex), arrayIndex,
-                    "Array index is beyond the end of the destination array.");
-            if (array.Length - arrayIndex < _dict._count)
-                throw new ArgumentException(
-                    "The destination array has insufficient space for the values.", nameof(array));
+            CopyToGuard.Validate(array, arrayIndex, _dict._count, CopyToGuard.ValuesMessage);
 
             int i = arrayIndex;
             foreach (KeyValuePair<TKey, TValue?> entry in _dict)